requirements.txt
mkdocs==1.6.0
mkdocs-material==9.5.26
mkdocs-bibtex==2.14.5
mkdocs-macros-plugin==1.0.5
mkdocs-awesome-pages-plugin==2.9.2
mkdocs-glightbox==0.4.0
mkdocs-redirects==1.2.1
mkdocs-section-index==0.3.9
pymdown-extensions==10.10.1
python-frontmatter==1.0.1
requests==2.32.3
PyYAML==6.0.2
tabulate==0.9.0
